requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
jinja2
python-dotenv
openai
